added file io funcs to separate file
